% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_text}
\alias{clean_text}
\title{Simple text cleaning and stopword removal}
\usage{
clean_text(text, stop_words = NULL, remove = NULL, case = "tolower")
}
\arguments{
\item{text}{character vector of text / documents}

\item{stop_words}{character vector of additional words removed from text}

\item{remove}{character vector of non-standard characters to be replaced with a space}

\item{case}{character one of "tolower" (default) or "toupper"}
}
\value{
character vector of text / documents
}
\description{

}
