% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ozunconf18_tweets}
\alias{ozunconf18_tweets}
\title{ozunconf18 tweets.}
\format{A data frame with 1000 rows and 16 variables:
\describe{
  \item{text}{content of tweet}
  \item{favorited}{is the tweet favourited}
  \item{favoriteCount}{number of likes}
  \item{created}{date of the tweet}
  \item{truncated}{is the tweet truncated}
  \item{id}{tweet id}
  ...
}}
\source{
\url{http://twitter.com}
}
\usage{
ozunconf18_tweets
}
\description{
A dataset containing recent tweets including the #ozunconf18 hashtag.
Data retrieved 22nd November 2018 using code:
\code{
rstats_tweets <- searchTwitter('#ozunconf18', n=1000) %>%
    twListToDF()
}
}
\keyword{datasets}
