% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rstats_tweets}
\alias{rstats_tweets}
\title{1000 rstats tweets.}
\format{A data frame with 1000 rows and 16 variables:
\describe{
  \item{text}{content of tweet}
  \item{favorited}{is the tweet favourited}
  \item{favoriteCount}{number of likes}
  \item{created}{date of the tweet}
  \item{truncated}{is the tweet truncated}
  \item{id}{tweet id}
  ...
}}
\source{
\url{http://twitter.com}
}
\usage{
rstats_tweets
}
\description{
A dataset containing the 1000 recent tweets including the #rstats hashtag.
Data retrieved 23rd November 2018 using code:
\code{
rstats_tweets <- searchTwitter('#rstats', n=1000) %>%
    twListToDF()
}
}
\keyword{datasets}
