% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termtable.R
\name{term_table}
\alias{term_table}
\title{Convert text to a term table}
\usage{
term_table(text, n = 5)
}
\arguments{
\item{text}{character vector of text / documents}

\item{n}{integer exclude terms with cont less than n}
}
\value{
data.frame
}
\description{
Convert text to a term table
}
